requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
requests==2.32.5
python-dotenv==1.2.2
gunicorn==21.2.0
